refactor: update imports in Chat and LearningPageClient components; remove unused dashboard data query
