+++ src/services/student/studentApi.ts
@@ -1,10 +1,7 @@
-import { use } from "react";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import type {
   CourseSections,
-  CourseStats,
   PaginatedCourses,
-  ActivityFeedResponse,
   DashboardData,
   DashboardStats,
   RecentActivity,
@@ -16,14 +13,10 @@ import type {
   PaginatedReviews,
   UpdateReviewRequest,
   UpdateReviewResponse,
-  Activity,
   PaginatedQuizResults,
   QuizResultDetails,
-  QuizSubmissionRequest,
   QuizSubmissionResponse,
-  DiscountUsage,
   PaginatedDiscountUsages,
-  AffiliatePayout,
   PaginatedAffiliatePayouts,
   AffiliatePayoutStats,
 } from "@/types/student/index";
@@ -143,79 +136,6 @@ export const studentApi = createApi({
       transformResponse: (response: { data: CourseSections }) => {
         return response.data;
       },
-    }),
-
-    // Get complete dashboard data with detailed activities
-    getDashboardDataComplete: builder.query<
-      DashboardData,
-      { page?: number; size?: number }
-    >({
-      async queryFn(
-        { page = 0, size = 10 },
-        _queryApi,
-        _extraOptions,
-        fetchWithBQ
-      ) {
-        try {
-          // Fetch enrolled courses
-          const coursesResult = await fetchWithBQ({
-            url: "/student/courses",
-            method: "GET",
-          });
-
-          if (coursesResult.error) {
-            return { error: coursesResult.error };
-          }
-
-          const coursesData = extractApiData<PaginatedCourses>(
-            coursesResult as any
-          );
-          const courses = coursesData?.content ?? [];
-
-          // Return empty data if no courses found
-          if (courses.length === 0) {
-            return { data: getEmptyDashboardData(size) };
-          }
-
-          // Fetch sections for all courses with error handling
-          const courseWithSections = await fetchCourseSections(
-            courses,
-            fetchWithBQ
-          );
-
-          // Calculate lesson statistics
-          const lessonStats = calculateLessonStatistics(courseWithSections);
-
-          // Calculate course statistics
-          const stats = calculateCourseStatistics(courses, lessonStats);
-
-          // Generate all activities (enrollment, lessons, quizzes)
-          const activities = generateAllActivities(courses, courseWithSections);
-
-          // Sort activities by date and create paginated response
-          const sortedActivities = sortActivitiesByDate(activities);
-          const paginatedActivities = createPaginatedActivities(
-            sortedActivities,
-            page,
-            size
-          );
-
-          return {
-            data: {
-              stats,
-              activities: paginatedActivities,
-            },
-          };
-        } catch (error) {
-          return {
-            error: {
-              status: "FETCH_ERROR",
-              error: error instanceof Error ? error.message : String(error),
-            },
-          };
-        }
-      },
-      providesTags: ["Course", "Lesson"],
     }),
 
     // Complete a lesson
@@ -506,7 +426,6 @@ export const {
   useGetAllEnrolledCoursesQuery,
   useGetCourseDetailsQuery,
   useGetCourseSectionsQuery,
-  useGetDashboardDataCompleteQuery,
   useGetDashboardStatsQuery,
   useGetRecentActivitiesQuery,
   useCompleteLessonMutation,
